test: no-ticket: Split large OutOfBand test classes for better fork packing.
With three test workers, the testOutOfBand job's wall time is set by how
well classes pack into forks; single large classes scheduled late extend
the job past the serial/worker floor.

- Split BarrageMessageRoundTripTest (9.6 min, scheduled last in nightly
  runs) into an abstract harness base plus three category-annotated
  classes: subscription/viewport changes, incremental updates, and
  column-type/dictionary round trips.
- Move testSumByIncremental (two thirds of QueryTableAggregationTest's
  runtime) into QueryTableSumByIncrementalTest.
- Reduce RspBitmapTimeDrivenTest's default per-test budget from two
  minutes to one, cutting its fixed cost from six minutes to three.
