--- server/src/test/java/io/deephaven/server/barrage/BarrageMessageRoundTripTest.java
@@ -1,2479 +1,0 @@
-//
-// Copyright (c) 2016-2026 Deephaven Data Labs and Patent Pending
-//
-package io.deephaven.server.barrage;
-
-import com.google.protobuf.CodedInputStream;
-import dagger.BindsInstance;
-import dagger.Component;
-import io.deephaven.api.ColumnName;
-import io.deephaven.api.RawString;
-import io.deephaven.api.Selectable;
-import io.deephaven.base.Pair;
-import io.deephaven.base.verify.Assert;
-import io.deephaven.client.impl.BarrageSubscriptionImpl.BarrageDataMarshaller;
-import io.deephaven.engine.context.ExecutionContext;
-import io.deephaven.engine.rowset.*;
-import io.deephaven.engine.table.ModifiedColumnSet;
-import io.deephaven.engine.table.Table;
-import io.deephaven.engine.table.TableDefinition;
-import io.deephaven.engine.table.TableUpdate;
-import io.deephaven.engine.table.impl.InstrumentedTableUpdateListener;
-import io.deephaven.engine.table.impl.QueryTable;
-import io.deephaven.engine.table.impl.TableUpdateImpl;
-import io.deephaven.engine.table.impl.TableUpdateValidator;
-import io.deephaven.engine.table.impl.util.BarrageMessage;
-import io.deephaven.engine.table.vectors.IntVectorColumnWrapper;
-import io.deephaven.engine.testutil.*;
-import io.deephaven.engine.testutil.generator.*;
-import io.deephaven.engine.testutil.testcase.RefreshingTableTestCase;
-import io.deephaven.engine.updategraph.UpdateSourceCombiner;
-import io.deephaven.engine.util.TableDiff;
-import io.deephaven.engine.util.TableTools;
-import io.deephaven.extensions.barrage.BarrageMessageWriter;
-import io.deephaven.extensions.barrage.BarrageSubscriptionOptions;
-import io.deephaven.extensions.barrage.table.BarrageTable;
-import io.deephaven.extensions.barrage.util.BarrageMessageReaderImpl;
-import io.deephaven.extensions.barrage.util.BarrageProtoUtil;
-import io.deephaven.extensions.barrage.util.BarrageUtil;
-import io.deephaven.extensions.barrage.util.ExposedByteArrayOutputStream;
-import io.deephaven.extensions.barrage.util.GrpcMarshallingException;
-import io.deephaven.server.arrow.ArrowModule;
-import io.deephaven.server.session.SessionService;
-import io.deephaven.server.util.Scheduler;
-import io.deephaven.server.util.TestControlledScheduler;
-import io.deephaven.test.types.OutOfBandTest;
-import io.deephaven.time.DateTimeUtils;
-import io.deephaven.util.annotations.ReferentialIntegrity;
-import io.deephaven.util.annotations.ScriptApi;
-import io.deephaven.util.mutable.MutableInt;
-import io.deephaven.vector.IntVector;
-import io.grpc.StatusRuntimeException;
-import io.grpc.stub.StreamObserver;
-import junit.framework.TestCase;
-import org.apache.commons.lang3.mutable.MutableObject;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.junit.experimental.categories.Category;
-
-import javax.inject.Singleton;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.nio.ByteBuffer;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.ZoneId;
-import java.util.*;
-import java.util.function.Consumer;
-import java.util.function.Supplier;
-import java.util.function.UnaryOperator;
-import org.apache.arrow.flatbuf.Message;
-import org.apache.arrow.flatbuf.MessageHeader;
-import org.apache.arrow.vector.types.pojo.ArrowType;
-import org.apache.arrow.vector.types.pojo.DictionaryEncoding;
-import org.apache.arrow.vector.types.pojo.Field;
-import org.apache.arrow.vector.types.pojo.FieldType;
-import org.apache.arrow.vector.types.pojo.Schema;
-import java.util.stream.Collectors;
-
-import static io.deephaven.engine.table.impl.remote.ConstructSnapshot.SNAPSHOT_CHUNK_SIZE;
-import static io.deephaven.engine.testutil.TstUtils.*;
-import static io.deephaven.engine.util.TableTools.col;
-
-@Category(OutOfBandTest.class)
-public class BarrageMessageRoundTripTest extends RefreshingTableTestCase {
-    private static final long UPDATE_INTERVAL = 1000; // arbitrary; we enforce coalescing on both sides
-
-    private TestControlledScheduler scheduler;
-    private Deque<Throwable> exceptions;
-    private UpdateSourceCombiner updateSourceCombiner;
-    private boolean useDeephavenNulls;
-    /** Message readers retain dictionary value chunks; they must be closed before the leak check in tearDown. */
-    private List<BarrageMessageReaderImpl> openMessageReaders;
-
-    private TestComponent daggerRoot;
-
-    @Singleton
-    @Component(modules = {
-            ArrowModule.class
-    })
-    public interface TestComponent {
-        BarrageMessageWriter.Factory getStreamGeneratorFactory();
-
-        @Component.Builder
-        interface Builder {
-            @BindsInstance
-            Builder withScheduler(final Scheduler scheduler);
-
-            TestComponent build();
-        }
-    }
-
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
-        updateSourceCombiner = new UpdateSourceCombiner(ExecutionContext.getContext().getUpdateGraph());
-        scheduler = new TestControlledScheduler();
-        exceptions = new ArrayDeque<>();
-        useDeephavenNulls = true;
-        openMessageReaders = new ArrayList<>();
-
-        daggerRoot = DaggerBarrageMessageRoundTripTest_TestComponent
-                .builder()
-                .withScheduler(scheduler)
-                .build();
-    }
-
-    @Override
-    public void tearDown() throws Exception {
-        openMessageReaders.forEach(BarrageMessageReaderImpl::close);
-        openMessageReaders = null;
-        updateSourceCombiner = null;
-        scheduler = null;
-        exceptions = null;
-        super.tearDown();
-    }
-
-    private void flushProducerTable() {
-        scheduler.runUntilQueueEmpty();
-    }
-
-    // We should listen for failures on the table, and if we get any, the test case is no good.
-    class FailureListener extends InstrumentedTableUpdateListener {
-        final String tableName;
-
-        FailureListener(String tableName) {
-            super("Failure Listener");
-            this.tableName = tableName;
-        }
-
-        @Override
-        public void onUpdate(final TableUpdate upstream) {
-            if (RefreshingTableTestCase.printTableUpdates) {
-                System.out.println("Incremental Table Update: (" + tableName + ")");
-                System.out.println(upstream);
-            }
-        }
-
-        @Override
-        public void onFailureInternal(Throwable originalException, Entry sourceEntry) {
-            exceptions.add(originalException);
-            final StringWriter errors = new StringWriter();
-            originalException.printStackTrace(new PrintWriter(errors));
-            TestCase.fail(errors.toString());
-        }
-    }
-
-    private class RemoteClient {
-        private RowSet viewport;
-        private boolean reverseViewport;
-
-        private BitSet subscribedColumns;
-
-        private final String name;
-
-        private final BarrageTable barrageTable;
-        @ReferentialIntegrity
-        private final BarrageMessageProducer barrageMessageProducer;
-
-        /** whether this is a full subscription, which {@link #viewport} does not record once it is changed */
-        private final boolean isFullSubscription;
-        /** set once the server has acknowledged a full subscription, so the replicated table is complete */
-        private boolean fullSubscriptionSatisfied;
-
-        @ReferentialIntegrity
-        private TableUpdateValidator replicatedTUV;
-        @ReferentialIntegrity
-        private FailureListener replicatedTUVListener;
-
-        private boolean subscribed = false;
-        private final Queue<BarrageMessage> commandQueue = new ArrayDeque<>();
-        private final DummyObserver dummyObserver;
-
-        private UnaryOperator<QueryTable> transform = null;
-
-        // The replicated table's TableUpdateValidator will be confused if the table is a viewport. Instead we rely on
-        // comparing the producer table to the consumer table to validate contents are correct.
-        RemoteClient(final RowSet viewport, final BitSet subscribedColumns,
-                final BarrageMessageProducer barrageMessageProducer,
-                final Table sourceTable, final String name) {
-            // assume a forward viewport when not specified
-            this(viewport, subscribedColumns, barrageMessageProducer, sourceTable, name, false, false);
-        }
-
-        RemoteClient(final RowSet viewport, final BitSet subscribedColumns,
-                final BarrageMessageProducer barrageMessageProducer,
-                final Table sourceTable,
-                final String name, final boolean reverseViewport, final boolean deferSubscription) {
-            this.viewport = viewport;
-            this.reverseViewport = reverseViewport;
-            this.subscribedColumns = subscribedColumns;
-            this.name = name;
-            this.barrageMessageProducer = barrageMessageProducer;
-            this.isFullSubscription = viewport == null;
-
-            final Map<String, Object> attributes = new HashMap<>(sourceTable.getAttributes());
-            if (sourceTable.isFlat()) {
-                attributes.put(BarrageUtil.TABLE_ATTRIBUTE_IS_FLAT, true);
-            }
-            final BarrageUtil.ConvertedArrowSchema schema = BarrageUtil.convertArrowSchema(BarrageUtil.makeSchema(
-                    BarrageUtil.DEFAULT_SNAPSHOT_OPTIONS, barrageMessageProducer.getTableDefinition(), attributes,
-                    sourceTable.isFlat()));
-            this.barrageTable =
-                    BarrageTable.make(null, updateSourceCombiner, ExecutionContext.getContext().getUpdateGraph(),
-                            null, schema, isFullSubscription, new BarrageTable.ViewportChangedCallback() {
-                                @Override
-                                public boolean viewportChanged(
-                                        @Nullable final RowSet rowSet,
-                                        @Nullable final BitSet columns,
-                                        final boolean reverse) {
-                                    if (!isFullSubscription || rowSet != null) {
-                                        // the server is still growing this subscription toward the entire table
-                                        return true;
-                                    }
-                                    fullSubscriptionSatisfied = true;
-                                    return false;
-                                }
-
-                                @Override
-                                public void onError(@NotNull final Throwable t) {
-                                    exceptions.add(t);
-                                }
-                            });
-            this.barrageTable.addSourceToRegistrar();
-
-            final BarrageSubscriptionOptions options = BarrageSubscriptionOptions.builder()
-                    .useDeephavenNulls(useDeephavenNulls)
-                    .build();
-            final BarrageMessageReaderImpl messageReader =
-                    new BarrageMessageReaderImpl(barrageTable.getDeserializationTmConsumer());
-            openMessageReaders.add(messageReader);
-            final BarrageDataMarshaller marshaller = new BarrageDataMarshaller(
-                    options, schema.computeWireChunkTypes(), schema.computeWireTypes(),
-                    schema.computeWireComponentTypes(), messageReader);
-            this.dummyObserver = new DummyObserver(marshaller, commandQueue);
-
-            // Note that no TableUpdateValidator is created for a viewport subscription: the TUV is unaware of the
-            // viewport and gets confused about which data should be valid. Instead we rely on the validation of the
-            // content in the viewport between the consumer and expected table. A full subscription's validator is
-            // attached by maybeAttachUpdateValidator() once the replicated table is complete.
-
-            if (!deferSubscription) {
-                doSubscribe();
-            }
-        }
-
-        public void setTransform(UnaryOperator<QueryTable> transform) {
-            this.transform = transform;
-        }
-
-        public void doSubscribe() {
-            subscribed = true;
-            final BarrageSubscriptionOptions options = BarrageSubscriptionOptions.builder()
-                    .useDeephavenNulls(useDeephavenNulls)
-                    .build();
-            barrageMessageProducer.addSubscription(dummyObserver, options, subscribedColumns,
-                    viewport == null ? null : viewport.copy(), reverseViewport);
-        }
-
-        /**
-         * Attaches the replicated table's {@link TableUpdateValidator}, but not before the server has acknowledged the
-         * full subscription. Until then the replicated table is incomplete: the server is still growing the client's
-         * viewport toward the entire table and publishes no updates for the rows it delivers, so a validator attached
-         * any earlier would be validating a table whose initial contents it was never given.
-         */
-        private void maybeAttachUpdateValidator() {
-            if (replicatedTUV != null || !fullSubscriptionSatisfied) {
-                return;
-            }
-            replicatedTUV = TableUpdateValidator.make(barrageTable);
-            replicatedTUVListener = new FailureListener("Replicated Table Update Validator");
-            replicatedTUV.getResultTable().addUpdateListener(replicatedTUVListener);
-        }
-
-        public void validate(final String msg, QueryTable expected) {
-            if (!subscribed) {
-                return; // no subscription implies no run implies no data -- so we're valid
-            }
-
-            maybeAttachUpdateValidator();
-
-            // We expect all messages from original table to have been propagated to the replicated table at this point.
-
-            QueryTable toCheck = barrageTable;
-            if (viewport != null) {
-                expected = expected
-                        .getSubTable(expected.getRowSet().subSetForPositions(viewport, reverseViewport).toTracking());
-            }
-            if (subscribedColumns.cardinality() != expected.numColumns()) {
-                final List<Selectable> columns = new ArrayList<>();
-                for (int i = subscribedColumns.nextSetBit(0); i >= 0; i = subscribedColumns.nextSetBit(i + 1)) {
-                    columns.add(ColumnName.of(expected.getDefinition().getColumns().get(i).getName()));
-                }
-                expected = (QueryTable) expected.view(columns);
-                toCheck = (QueryTable) toCheck.view(columns);
-            }
-
-            if (transform != null) {
-                expected = transform.apply(expected);
-                toCheck = transform.apply(toCheck);
-            }
-
-            // Data should be identical and in-order.
-            TstUtils.assertTableEquals(expected, toCheck);
-            if (viewport == null) {
-                // Since key-space needs to be kept the same, the RowSets should also be identical between producer and
-                // consumer (not RowSets between expected and consumer; as the consumer maintains the entire RowSet).
-                Assert.equals(barrageMessageProducer.getRowSet(), "barrageMessageProducer.getRowSet()",
-                        barrageTable.getRowSet(), "barrageTable.getRowSet()");
-            } else {
-                // otherwise, the RowSet should represent a flattened view of the viewport
-                Assert.eqTrue(barrageTable.getRowSet().isFlat(), "barrageTable.getRowSet().isFlat()");
-            }
-        }
-
-        private void showResult(final String label, final Table table) {
-            System.out.println(label);
-            TableTools.showWithRowSet(table, 100);
-        }
-
-        public void show(QueryTable expected) {
-            QueryTable toCheck = barrageTable;
-            if (viewport != null) {
-                expected = expected.getSubTable(expected.getRowSet().subSetForPositions(viewport).toTracking());
-                toCheck = toCheck.getSubTable(toCheck.getRowSet().subSetForPositions(viewport).toTracking());
-            }
-            if (subscribedColumns.cardinality() != expected.numColumns()) {
-                final List<Selectable> columns = new ArrayList<>();
-                for (int i = subscribedColumns.nextSetBit(0); i >= 0; i = subscribedColumns.nextSetBit(i + 1)) {
-                    columns.add(ColumnName.of(expected.getDefinition().getColumns().get(i).getName()));
-                }
-                expected = (QueryTable) expected.view(columns);
-                toCheck = (QueryTable) toCheck.view(columns);
-            }
-
-            final int maxLines = 100;
-            final Pair<String, Long> diffPair =
-                    TableTools.diffPair(toCheck, expected, maxLines, EnumSet.of(TableDiff.DiffItems.DoublesExact));
-
-            if (diffPair.getFirst().equals("")) {
-                showResult("Ticking Table (" + name + "):", toCheck);
-            } else {
-                final long numTableRows = Math.min(maxLines, Math.max(toCheck.size(), expected.size()));
-                final long firstRow = Math.max(0, diffPair.getSecond() - 5);
-                final long lastRow =
-                        Math.min(firstRow + numTableRows, Math.min(firstRow + maxLines, diffPair.getSecond() + 5));
-
-                System.out.println("Recomputed Table (" + name + ") Differs:\n" + diffPair.getFirst()
-                        + "\nRecomputed Table Rows [" + firstRow + ", " + lastRow + "]:");
-                TableTools.showWithRowSet(expected, firstRow, lastRow + 1);
-                System.out.println("Replicated Table Rows [" + firstRow + ", " + lastRow + "]:");
-                TableTools.showWithRowSet(toCheck, firstRow, lastRow + 1);
-            }
-        }
-
-        public void flushEventsToReplicatedTable() {
-            for (final BarrageMessage msg : commandQueue) {
-                barrageTable.handleBarrageMessage(msg);
-                msg.close();
-            }
-            commandQueue.clear();
-        }
-
-        public int pendingMessageCount() {
-            return commandQueue.size();
-        }
-
-        /**
-         * Delivers at most {@code maxMessages} of the queued messages, leaving the rest for a later flush. Callers use
-         * this to spread a single server propagation across more than one consumer update graph cycle.
-         */
-        public void flushEventsToReplicatedTable(final int maxMessages) {
-            for (int ii = 0; ii < maxMessages; ++ii) {
-                final BarrageMessage msg = commandQueue.poll();
-                if (msg == null) {
-                    return;
-                }
-                barrageTable.handleBarrageMessage(msg);
-                msg.close();
-            }
-        }
-
-        public void setViewport(final RowSet newViewport) {
-            // assume a forward viewport when not specified
-            setViewport(newViewport, false);
-        }
-
-        public void setViewport(final RowSet newViewport, final boolean newReverseViewport) {
-            viewport = newViewport;
-            reverseViewport = newReverseViewport;
-
-            // maintain the existing subscribedColumns set
-            barrageMessageProducer.updateSubscription(dummyObserver, viewport, subscribedColumns, reverseViewport);
-        }
-
-        public void setSubscribedColumns(final BitSet newColumns) {
-            subscribedColumns = newColumns;
-
-            // maintain the existing viewport and viewport direction
-            barrageMessageProducer.updateSubscription(dummyObserver, viewport, subscribedColumns, reverseViewport);
-        }
-
-        public void setViewportAndColumns(final RowSet newViewport, final BitSet newColumns) {
-            // assume a forward viewport when not specified
-            setViewportAndColumns(newViewport, newColumns, false);
-        }
-
-        public void setViewportAndColumns(final RowSet newViewport, final BitSet newColumns,
-                final boolean newReverseViewport) {
-            viewport = newViewport;
-            reverseViewport = newReverseViewport;
-            subscribedColumns = newColumns;
-            barrageMessageProducer.updateSubscription(dummyObserver, viewport, subscribedColumns, newReverseViewport);
-        }
-    }
-
-    private class RemoteNugget implements EvalNuggetInterface {
-
-        private final Supplier<Table> makeTable;
-
-        private final QueryTable originalTable;
-        @ReferentialIntegrity
-        private final BarrageMessageProducer barrageMessageProducer;
-
-        @ReferentialIntegrity
-        private final TableUpdateValidator originalTUV;
-        @ReferentialIntegrity
-        private final FailureListener originalTUVListener;
-
-        private final List<RemoteClient> clients = new ArrayList<>();
-
-        RemoteNugget(final Supplier<Table> makeTable) {
-            this.makeTable = makeTable;
-            this.originalTable = (QueryTable) makeTable.get();
-            this.barrageMessageProducer = originalTable.getResult(new BarrageMessageProducer.Operation(scheduler,
-                    new SessionService.ObfuscatingErrorTransformer(), daggerRoot.getStreamGeneratorFactory(),
-                    originalTable, UPDATE_INTERVAL, this::onGetSnapshot));
-
-            originalTUV = TableUpdateValidator.make(originalTable);
-            originalTUVListener = new FailureListener("Original Table Update Validator");
-            originalTUV.getResultTable().addUpdateListener(originalTUVListener);
-        }
-
-        @Override
-        public void validate(final String msg) {
-            final QueryTable expected = (QueryTable) makeTable.get();
-            for (final RemoteClient client : clients) {
-                client.validate(msg, expected);
-            }
-        }
-
-        @Override
-        public void show() {
-            final QueryTable expected = (QueryTable) makeTable.get();
-            for (final RemoteClient client : clients) {
-                client.show(expected);
-            }
-        }
-
-        public void flushClientEvents() {
-            for (final RemoteClient client : clients) {
-                client.flushEventsToReplicatedTable();
-            }
-        }
-
-        public RemoteClient newClient(final RowSet viewport, final BitSet subscribedColumns, final String name) {
-            // assume a forward viewport when not specified
-            return newClient(viewport, subscribedColumns, false, name);
-        }
-
-        public RemoteClient newClient(final RowSet viewport, final BitSet subscribedColumns,
-                final boolean reverseViewport, final String name) {
-            clients.add(new RemoteClient(viewport, subscribedColumns, barrageMessageProducer,
-                    originalTable, name, reverseViewport, false));
-            return clients.get(clients.size() - 1);
-        }
-
-        public void onGetSnapshot() {}
-    }
-
-    private abstract class TestHelper {
-        final int numProducerCoalesce;
-        final int numConsumerCoalesce;
-
-        final int size;
-        final Random random;
-        final MutableInt numSteps;
-
-        final List<RemoteNugget> nuggets = new ArrayList<>();
-
-        QueryTable sourceTable;
-        ColumnInfo<?, ?>[] columnInfo;
-
-        TestHelper(final int numProducerCoalesce, final int numConsumerCoalesce, final int size, final int seed,
-                final MutableInt numSteps) {
-            this.numProducerCoalesce = numProducerCoalesce;
-            this.numConsumerCoalesce = numConsumerCoalesce;
-            this.size = size;
-            this.random = new Random(seed);
-            this.numSteps = numSteps;
-        }
-
-        public void createTable() {
-            sourceTable = getTable(size / 4, random,
-                    columnInfo =
-                            initColumnInfos(
-                                    new String[] {"Sym", "intCol", "doubleCol", "Indices", "boolCol", "TimeStamp"},
-                                    new SetGenerator<>("a", "b", "c", "d"),
-                                    new IntGenerator(10, 100),
-                                    new SetGenerator<>(10.1, 20.1, 30.1),
-                                    new SortedLongGenerator(0, Long.MAX_VALUE - 1),
-                                    new BooleanGenerator(0.2),
-                                    new UnsortedInstantGenerator(
-                                            DateTimeUtils.parseInstant("2020-02-14T00:00:00 NY"),
-                                            DateTimeUtils.parseInstant("2020-02-25T00:00:00 NY"))));
-        }
-
-        public void createNuggets() {
-            // test the explicit updates
-            createNuggetsForTableMaker(() -> sourceTable);
-            // test shift aggressive version of these updates
-            createNuggetsForTableMaker(sourceTable::flatten);
-            // test updates in the middle of the keyspace
-            createNuggetsForTableMaker(() -> sourceTable.sort("doubleCol"));
-            // test sparse(r) updates
-            createNuggetsForTableMaker(() -> sourceTable.where("intCol % 12 < 5"));
-            // test for the nested Vector encoding/decoding (though most types are tested
-            createNuggetsForTableMaker(() -> sourceTable.groupBy("Sym").sort("Sym"));
-        }
-
-        void runTest(final Runnable simulateSourceStep) {
-            createTable();
-            createNuggets();
-            final int maxSteps = numSteps.get();
-            final RemoteNugget[] nuggetsToValidate = nuggets.toArray(new RemoteNugget[0]);
-            for (numSteps.set(0); numSteps.get() < maxSteps; numSteps.increment()) {
-                for (int rt = 0; rt < numConsumerCoalesce; ++rt) {
-                    // coalesce updates in producer
-                    for (int pt = 0; pt < numProducerCoalesce; ++pt) {
-                        simulateSourceStep.run();
-                    }
-
-                    flushProducerTable();
-                }
-
-                // flush consumer
-                for (final RemoteNugget nugget : nuggets) {
-                    nugget.flushClientEvents();
-                }
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-
-                TstUtils.validate("", nuggetsToValidate);
-
-                if (sourceTable.size() >= size) {
-                    numSteps.set(maxSteps); // pretend we finished
-                    return;
-                }
-            }
-        }
-
-        abstract void createNuggetsForTableMaker(final Supplier<Table> makeTable);
-    }
-
-    private class OneProducerPerClient extends TestHelper {
-        OneProducerPerClient(final int numProducerCoalesce, final int numConsumerCoalesce, final int size,
-                final int seed, final MutableInt numSteps) {
-            super(numProducerCoalesce, numConsumerCoalesce, size, seed, numSteps);
-        }
-
-        void createNuggetsForTableMaker(final Supplier<Table> makeTable) {
-            final BitSet subscribedColumns = new BitSet();
-            subscribedColumns.set(0, makeTable.get().numColumns());
-
-            nuggets.add(new RemoteNugget(makeTable));
-            nuggets.get(nuggets.size() - 1).newClient(null, subscribedColumns, "full");
-
-            nuggets.add(new RemoteNugget(makeTable));
-            nuggets.get(nuggets.size() - 1).newClient(RowSetFactory.fromRange(0, size / 10),
-                    subscribedColumns, "header");
-            nuggets.add(new RemoteNugget(makeTable));
-            nuggets.get(nuggets.size() - 1).newClient(
-                    RowSetFactory.fromRange(size / 2, size * 3L / 4),
-                    subscribedColumns, "floating");
-
-            nuggets.add(new RemoteNugget(makeTable));
-            nuggets.get(nuggets.size() - 1).newClient(
-                    RowSetFactory.fromRange(0, size / 10),
-                    subscribedColumns, true, "footer");
-            nuggets.add(new RemoteNugget(makeTable));
-            nuggets.get(nuggets.size() - 1).newClient(
-                    RowSetFactory.fromRange(size / 2, size * 3L / 4),
-                    subscribedColumns, true, "reverse floating");
-
-            final RowSetBuilderSequential swissIndexBuilder = RowSetFactory.builderSequential();
-            final long rangeSize = Math.max(1, size / 20);
-            for (long nr = 1; nr < 20; nr += 2) {
-                swissIndexBuilder.appendRange(nr * rangeSize, (nr + 1) * rangeSize - 1);
-            }
-            final RowSet rs = swissIndexBuilder.build();
-
-            nuggets.add(new RemoteNugget(makeTable));
-            nuggets.get(nuggets.size() - 1).newClient(rs, subscribedColumns, "swiss");
-
-
-            final RemoteNugget nugget = new RemoteNugget(makeTable);
-            nugget.newClient(rs.copy(), subscribedColumns, true, "reverse swiss");
-            nuggets.add(nugget);
-        }
-    }
-
-    private class SharedProducerForAllClients extends TestHelper {
-        SharedProducerForAllClients(final int numProducerCoalesce, final int numConsumerCoalesce, final int size,
-                final int seed, final MutableInt numSteps) {
-            super(numProducerCoalesce, numConsumerCoalesce, size, seed, numSteps);
-        }
-
-        void createNuggetsForTableMaker(final Supplier<Table> makeTable) {
-            final BitSet subscribedColumns = new BitSet();
-            subscribedColumns.set(0, makeTable.get().numColumns());
-
-            final RemoteNugget nugget = new RemoteNugget(makeTable);
-            nuggets.add(nugget);
-            nugget.newClient(null, subscribedColumns, "full");
-
-            nugget.newClient(RowSetFactory.fromRange(0, size / 10), subscribedColumns, "header");
-            nugget.newClient(RowSetFactory.fromRange(size / 2, size * 3L / 4), subscribedColumns, "floating");
-
-            nugget.newClient(RowSetFactory.fromRange(0, size / 10), subscribedColumns, true, "footer");
-            nugget.newClient(RowSetFactory.fromRange(size / 2, size * 3L / 4), subscribedColumns, true,
-                    "reverse floating");
-
-            final RowSetBuilderSequential swissIndexBuilder = RowSetFactory.builderSequential();
-            final long rangeSize = Math.max(1, size / 20);
-            for (long nr = 1; nr < 20; nr += 2) {
-                swissIndexBuilder.appendRange(nr * rangeSize, (nr + 1) * rangeSize - 1);
-            }
-
-            final RowSet rs = swissIndexBuilder.build();
-            nugget.newClient(rs, subscribedColumns, "swiss");
-
-            nugget.newClient(rs.copy(), subscribedColumns, true, "reverse swiss");
-        }
-    }
-
-    public void testAppendIncremental() {
-        final int MAX_STEPS = 100;
-        final Consumer<TestHelper> runOne = helper -> {
-            final int maxSteps = MAX_STEPS * helper.numConsumerCoalesce * helper.numProducerCoalesce;
-            helper.runTest(() -> {
-                final long lastKey = (Math.abs(helper.random.nextLong()) % 16)
-                        + (helper.sourceTable.getRowSet().isNonempty() ? helper.sourceTable.getRowSet().lastRowKey()
-                                : -1);
-                final TableUpdateImpl update = new TableUpdateImpl();
-                update.added = RowSetFactory.fromRange(lastKey + 1,
-                        lastKey + Math.max(1, helper.size / maxSteps));
-                update.removed = i();
-                update.modified = i();
-                update.shifted = RowSetShiftData.EMPTY;
-                update.modifiedColumnSet = ModifiedColumnSet.EMPTY;
-
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateTableUpdates(update,
-                        helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        };
-
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 10}) {
-                for (final int numConsumerCoalesce : new int[] {1, 10}) {
-                    runOne.accept(new OneProducerPerClient(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                            new MutableInt(MAX_STEPS)));
-                }
-            }
-        }
-    }
-
-    public void testPrependIncremental() {
-        final int MAX_STEPS = 100;
-        final Consumer<TestHelper> runOne = helper -> {
-            final int maxSteps = MAX_STEPS * helper.numConsumerCoalesce * helper.numProducerCoalesce;
-            helper.runTest(() -> {
-                final long lastKey =
-                        helper.sourceTable.getRowSet().isNonempty() ? helper.sourceTable.getRowSet().lastRowKey() : -1;
-                final TableUpdateImpl update = new TableUpdateImpl();
-                final int stepSize = Math.max(1, helper.size / maxSteps);
-                update.added = RowSetFactory.fromRange(0, stepSize - 1);
-                update.removed = i();
-                update.modified = i();
-                update.modifiedColumnSet = ModifiedColumnSet.EMPTY;
-
-                final RowSetShiftData.Builder shifted = new RowSetShiftData.Builder();
-                if (lastKey >= 0) {
-                    shifted.shiftRange(0, lastKey, stepSize + (Math.abs(helper.random.nextLong()) % 16));
-                }
-                update.shifted = shifted.build();
-
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateTableUpdates(update,
-                        helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        };
-
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 10}) {
-                for (final int numConsumerCoalesce : new int[] {1, 10}) {
-                    runOne.accept(new OneProducerPerClient(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                            new MutableInt(MAX_STEPS)));
-                }
-            }
-        }
-    }
-
-    public void testRoundTripIncremental() {
-        final Consumer<TestHelper> runOne = helper -> {
-            helper.runTest(() -> {
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                        GenerateTableUpdates.DEFAULT_PROFILE,
-                        helper.size, helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        };
-
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 10}) {
-                for (final int numConsumerCoalesce : new int[] {1, 10}) {
-                    runOne.accept(new OneProducerPerClient(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                            new MutableInt(100)));
-                }
-            }
-        }
-    }
-
-    public void testAppendIncrementalSharedProducer() {
-        final int MAX_STEPS = 100;
-        final Consumer<TestHelper> runOne = helper -> {
-            final int maxSteps = MAX_STEPS * helper.numConsumerCoalesce * helper.numProducerCoalesce;
-            helper.runTest(() -> {
-                final long lastKey = (Math.abs(helper.random.nextLong()) % 16)
-                        + (helper.sourceTable.getRowSet().isNonempty() ? helper.sourceTable.getRowSet().lastRowKey()
-                                : -1);
-                final TableUpdateImpl update = new TableUpdateImpl();
-                update.added = RowSetFactory.fromRange(lastKey + 1,
-                        lastKey + Math.max(1, helper.size / maxSteps));
-                update.removed = i();
-                update.modified = i();
-                update.shifted = RowSetShiftData.EMPTY;
-                update.modifiedColumnSet = ModifiedColumnSet.EMPTY;
-
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateTableUpdates(update,
-                        helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        };
-
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 2, 10}) {
-                for (final int numConsumerCoalesce : new int[] {1, 2, 10}) {
-                    runOne.accept(new SharedProducerForAllClients(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                            new MutableInt(MAX_STEPS)));
-                }
-            }
-        }
-    }
-
-    public void testPrependIncrementalSharedProducer() {
-        final int MAX_STEPS = 100;
-        final Consumer<TestHelper> runOne = helper -> {
-            final int maxSteps = MAX_STEPS * helper.numConsumerCoalesce * helper.numProducerCoalesce;
-            helper.runTest(() -> {
-                final long lastKey =
-                        helper.sourceTable.getRowSet().isNonempty() ? helper.sourceTable.getRowSet().lastRowKey() : -1;
-                final TableUpdateImpl update = new TableUpdateImpl();
-                final int stepSize = Math.max(1, helper.size / maxSteps);
-                update.added = RowSetFactory.fromRange(0, stepSize - 1);
-                update.removed = i();
-                update.modified = i();
-                update.modifiedColumnSet = ModifiedColumnSet.EMPTY;
-
-                final RowSetShiftData.Builder shifted = new RowSetShiftData.Builder();
-                if (lastKey >= 0) {
-                    shifted.shiftRange(0, lastKey, stepSize + (Math.abs(helper.random.nextLong()) % 16));
-                }
-                update.shifted = shifted.build();
-
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateTableUpdates(update,
-                        helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        };
-
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalsce : new int[] {1, 10}) {
-                for (final int numConsumerCoalesce : new int[] {1, 10}) {
-                    runOne.accept(new SharedProducerForAllClients(numProducerCoalsce, numConsumerCoalesce, size, 0,
-                            new MutableInt(MAX_STEPS)));
-                }
-            }
-        }
-    }
-
-    public void testRoundTripIncrementalSharedProducer() {
-        final Consumer<TestHelper> runOne = helper -> {
-            helper.runTest(() -> {
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                        GenerateTableUpdates.DEFAULT_PROFILE,
-                        helper.size, helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        };
-
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 10}) {
-                for (final int numConsumerCoalesce : new int[] {1, 10}) {
-                    runOne.accept(new SharedProducerForAllClients(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                            new MutableInt(100)));
-                }
-            }
-        }
-    }
-
-    // These test mid-cycle subscription changes and snapshot content
-    private abstract class SubscriptionChangingHelper extends SharedProducerForAllClients {
-        SubscriptionChangingHelper(final int numProducerCoalesce, final int numConsumerCoalesce, final int size,
-                final int seed, final MutableInt numSteps) {
-            super(numProducerCoalesce, numConsumerCoalesce, size, seed, numSteps);
-        }
-
-        void runTest() {
-            createTable();
-            createNuggets();
-            final int maxSteps = numSteps.get();
-            final RemoteNugget[] nuggetsToValidate = nuggets.toArray(new RemoteNugget[0]);
-            for (numSteps.set(0); numSteps.get() < maxSteps; numSteps.increment()) {
-                for (int rt = 0; rt < numConsumerCoalesce; ++rt) {
-                    // coalesce updates in producer
-                    for (int pt = 0; pt < numProducerCoalesce; ++pt) {
-                        maybeChangeSub(numSteps.get(), rt, pt);
-
-                        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                        updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                                GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-                    }
-
-                    // flush producer
-                    flushProducerTable();
-                }
-
-                // flush consumer
-                for (final RemoteNugget nugget : nuggets) {
-                    nugget.flushClientEvents();
-                }
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-
-                TstUtils.validate("", nuggetsToValidate);
-            }
-        }
-
-        abstract void maybeChangeSub(int step, int rt, int pt);
-    }
-
-    public void testColumnSubChange() {
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 4}) {
-                for (final int numConsumerCoalesce : new int[] {1, 4}) {
-                    for (int subProducerCoalesce =
-                            0; subProducerCoalesce < numProducerCoalesce; ++subProducerCoalesce) {
-                        for (int subConsumerCoalesce =
-                                0; subConsumerCoalesce < numConsumerCoalesce; ++subConsumerCoalesce) {
-                            final int finalSubProducerCoalesce = subProducerCoalesce;
-                            final int finalSubConsumerCoalesce = subConsumerCoalesce;
-                            new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                                    new MutableInt(4)) {
-                                {
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final BitSet columns = new BitSet();
-                                        columns.set(0, nugget.originalTable.numColumns() / 2);
-                                        nugget.clients.add(new RemoteClient(
-                                                RowSetFactory.fromRange(size / 5, 2L * size / 5),
-                                                columns, nugget.barrageMessageProducer, nugget.originalTable,
-                                                "sub-changer"));
-                                    }
-                                }
-
-                                void maybeChangeSub(final int step, final int rt, final int pt) {
-                                    if (step != 2 || rt != finalSubConsumerCoalesce || pt != finalSubProducerCoalesce) {
-                                        return;
-                                    }
-
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final RemoteClient client = nugget.clients.get(nugget.clients.size() - 1);
-                                        final BitSet columns = new BitSet();
-                                        final int numColumns = nugget.originalTable.numColumns();
-                                        columns.set(numColumns / 2, numColumns);
-                                        client.setSubscribedColumns(columns);
-                                    }
-                                }
-                            }.runTest();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    public void testViewportChange() {
-        for (final int size : new int[] {10, 100}) {
-            for (final int numProducerCoalesce : new int[] {1, 4}) {
-                for (final int numConsumerCoalesce : new int[] {1, 4}) {
-                    for (int subProducerCoalesce =
-                            0; subProducerCoalesce < numProducerCoalesce; ++subProducerCoalesce) {
-                        for (int subConsumerCoalesce =
-                                0; subConsumerCoalesce < numConsumerCoalesce; ++subConsumerCoalesce) {
-                            final int finalSubProducerCoalesce = 0;
-                            final int finalSubConsumerCoalesce = 1;
-                            new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                                    new MutableInt(25)) {
-                                @Override
-                                void createNuggetsForTableMaker(final Supplier<Table> makeTable) {
-                                    final RemoteNugget nugget = new RemoteNugget(makeTable);
-                                    nuggets.add(nugget);
-
-                                    final BitSet columns = new BitSet();
-                                    columns.set(0, 4);
-                                    nugget.clients.add(
-                                            new RemoteClient(RowSetFactory.fromRange(0, size / 5),
-                                                    columns, nugget.barrageMessageProducer, nugget.originalTable,
-                                                    "sub-changer"));
-                                }
-
-                                void maybeChangeSub(final int step, final int rt, final int pt) {
-                                    if (step % 2 != 0 || rt != finalSubConsumerCoalesce
-                                            || pt != finalSubProducerCoalesce) {
-                                        return;
-                                    }
-
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final RemoteClient client = nugget.clients.get(nugget.clients.size() - 1);
-                                        final WritableRowSet viewport = client.viewport.copy();
-                                        viewport.shiftInPlace(Math.max(size / 25, 1));
-
-                                        // maintain viewport direction in this test
-                                        client.setViewport(viewport, client.reverseViewport);
-                                    }
-                                }
-                            }.runTest();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    public void testViewportDirectionChange() {
-        for (final int size : new int[] {10, 100}) {
-            for (final int numProducerCoalesce : new int[] {1, 4}) {
-                for (final int numConsumerCoalesce : new int[] {1, 4}) {
-                    for (int subProducerCoalesce =
-                            0; subProducerCoalesce < numProducerCoalesce; ++subProducerCoalesce) {
-                        for (int subConsumerCoalesce =
-                                0; subConsumerCoalesce < numConsumerCoalesce; ++subConsumerCoalesce) {
-                            final int finalSubProducerCoalesce = 0;
-                            final int finalSubConsumerCoalesce = 1;
-                            new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                                    new MutableInt(25)) {
-                                @Override
-                                void createNuggetsForTableMaker(final Supplier<Table> makeTable) {
-                                    final RemoteNugget nugget = new RemoteNugget(makeTable);
-                                    nuggets.add(nugget);
-
-                                    final BitSet columns = new BitSet();
-                                    columns.set(0, 4);
-                                    nugget.clients.add(
-                                            new RemoteClient(RowSetFactory.fromRange(0, size / 5),
-                                                    columns, nugget.barrageMessageProducer, nugget.originalTable,
-                                                    "sub-changer"));
-                                }
-
-                                void maybeChangeSub(final int step, final int rt, final int pt) {
-                                    if (step % 2 != 0 || rt != finalSubConsumerCoalesce
-                                            || pt != finalSubProducerCoalesce) {
-                                        return;
-                                    }
-
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final RemoteClient client = nugget.clients.get(nugget.clients.size() - 1);
-                                        final WritableRowSet viewport = client.viewport.copy();
-                                        viewport.shiftInPlace(Math.max(size / 25, 1));
-
-                                        // alternate viewport direction with every call to this function
-                                        client.setViewport(viewport, !client.reverseViewport);
-                                    }
-                                }
-                            }.runTest();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    public void testOverlappedColumnSubsChange() {
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 4}) {
-                for (final int numConsumerCoalesce : new int[] {1, 4}) {
-                    for (int subProducerCoalesce =
-                            0; subProducerCoalesce < numProducerCoalesce; ++subProducerCoalesce) {
-                        for (int subConsumerCoalesce =
-                                0; subConsumerCoalesce < numConsumerCoalesce; ++subConsumerCoalesce) {
-                            final int finalSubProducerCoalesce = subProducerCoalesce;
-                            final int finalSubConsumerCoalesce = subConsumerCoalesce;
-                            new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                                    new MutableInt(4)) {
-                                {
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final BitSet columns = new BitSet();
-                                        columns.set(0, 3);
-                                        nugget.clients.add(new RemoteClient(
-                                                RowSetFactory.fromRange(size / 5, 2L * size / 5),
-                                                columns, nugget.barrageMessageProducer, nugget.originalTable,
-                                                "sub-changer"));
-                                    }
-                                }
-
-                                void maybeChangeSub(final int step, final int rt, final int pt) {
-                                    if (step != 2 || rt != finalSubConsumerCoalesce || pt != finalSubProducerCoalesce) {
-                                        return;
-                                    }
-
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final RemoteClient client = nugget.clients.get(nugget.clients.size() - 1);
-                                        final BitSet columns = new BitSet();
-                                        columns.set(1, 4);
-                                        client.setSubscribedColumns(columns);
-                                    }
-                                }
-                            }.runTest();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    public void testViewportSubscribeMidCycle() {
-        // This is a regression test for IDS-6392. It catches a race between when a subscription becomes active and
-        // when the viewport becomes active post-snapshot.
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {2, 3, 4}) {
-                for (final int numConsumerCoalesce : new int[] {1, 4}) {
-                    new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                            new MutableInt(4)) {
-
-                        void maybeChangeSub(final int step, final int rt, final int pt) {
-                            if (step != 0 || rt != 0 || pt != 1) {
-                                // Only subscribe after we have sent at least one update, but no need to subscribe
-                                // again.
-                                return;
-                            }
-
-                            nuggets.forEach((nugget) -> nugget.clients.forEach(RemoteClient::doSubscribe));
-                        }
-
-                        @Override
-                        void createNuggetsForTableMaker(final Supplier<Table> makeTable) {
-                            if (!nuggets.isEmpty()) {
-                                return; // we can only have a single nugget since they all share a single source table
-                            }
-
-                            final RemoteNugget nugget = new RemoteNugget(makeTable) {
-                                @Override
-                                public void onGetSnapshot() {
-                                    final ControlledUpdateGraph updateGraph =
-                                            ExecutionContext.getContext().getUpdateGraph().cast();
-                                    updateGraph.runWithinUnitTestCycle(
-                                            () -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                                                    GenerateTableUpdates.DEFAULT_PROFILE, size,
-                                                    random, sourceTable, columnInfo));
-                                }
-                            };
-                            nuggets.add(nugget);
-
-                            // we can only have viewport subscriptions or else we won't tickle the original bug
-                            final BitSet columns = new BitSet();
-                            columns.set(0, 4);
-                            final boolean deferSubscription = true;
-                            nugget.clients.add(new RemoteClient(
-                                    RowSetFactory.fromRange(size / 5, 2L * size / 5),
-                                    columns, nugget.barrageMessageProducer, nugget.originalTable,
-                                    "sub-changer", false, deferSubscription));
-
-                        }
-                    }.runTest();
-                }
-            }
-        }
-    }
-
-    public void testOverlappingViewportChange() {
-        for (final int size : new int[] {10, 100, 1000}) {
-            for (final int numProducerCoalesce : new int[] {1, 4}) {
-                for (final int numConsumerCoalesce : new int[] {1, 4}) {
-                    for (int subProducerCoalesce =
-                            0; subProducerCoalesce < numProducerCoalesce; ++subProducerCoalesce) {
-                        for (int subConsumerCoalesce =
-                                0; subConsumerCoalesce < numConsumerCoalesce; ++subConsumerCoalesce) {
-                            final int finalSubProducerCoalesce = subProducerCoalesce;
-                            final int finalSubConsumerCoalesce = subConsumerCoalesce;
-                            new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0,
-                                    new MutableInt(4)) {
-                                @Override
-                                public void createNuggets() {
-                                    super.createNuggets();
-
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final BitSet columns = new BitSet();
-                                        columns.set(0, 4);
-                                        nugget.clients.add(new RemoteClient(
-                                                RowSetFactory.fromRange(size / 5, 3L * size / 5),
-                                                columns, nugget.barrageMessageProducer, nugget.originalTable,
-                                                "sub-changer"));
-                                    }
-                                }
-
-                                void maybeChangeSub(final int step, final int rt, final int pt) {
-                                    if (step != 2 || rt != finalSubConsumerCoalesce || pt != finalSubProducerCoalesce) {
-                                        return;
-                                    }
-
-                                    for (final RemoteNugget nugget : nuggets) {
-                                        final RemoteClient client = nugget.clients.get(nugget.clients.size() - 1);
-                                        final WritableRowSet viewport = client.viewport.copy();
-                                        viewport.shiftInPlace(size / 5);
-
-                                        // maintain viewport direction in this test
-                                        client.setViewport(viewport, client.reverseViewport);
-                                    }
-                                }
-                            }.runTest();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    public void testSimultaneousSubscriptionChanges() {
-        for (final int size : new int[] {10, 100, 1000}) {
-            final int numProducerCoalesce = 8;
-            final int numConsumerCoalesce = 8;
-            for (int subConsumerCoalesce = 0; subConsumerCoalesce < numConsumerCoalesce; ++subConsumerCoalesce) {
-                final int finalSubConsumerCoalesce = subConsumerCoalesce;
-                new SubscriptionChangingHelper(numProducerCoalesce, numConsumerCoalesce, size, 0, new MutableInt(4)) {
-                    {
-                        for (final RemoteNugget nugget : nuggets) {
-                            final BitSet columns = new BitSet();
-                            columns.set(0, 4);
-                            nugget.clients.add(new RemoteClient(
-                                    RowSetFactory.fromRange(size / 5, 2L * size / 5),
-                                    columns, nugget.barrageMessageProducer, nugget.originalTable, "sub-changer"));
-                        }
-                    }
-
-                    void maybeChangeSub(final int step, final int rt, final int pt) {
-                        if (step != 2 || rt != finalSubConsumerCoalesce) {
-                            return;
-                        }
-
-                        for (final RemoteNugget nugget : nuggets) {
-                            final RemoteClient client = nugget.clients.get(nugget.clients.size() - 1);
-                            final int firstKey = random.nextInt(size);
-                            client.setViewport(RowSetFactory.fromRange(firstKey,
-                                    firstKey + random.nextInt(size - firstKey)), client.reverseViewport);
-                        }
-                    }
-                }.runTest();
-            }
-        }
-    }
-
-    public void testUsePrevOnSnapshot() {
-        final QueryTable queryTable = TstUtils.testRefreshingTable(i(10, 12).toTracking(), col("intCol", 10, 12));
-        final RemoteNugget remoteNugget = new RemoteNugget(() -> queryTable);
-        final MutableObject<RemoteClient> remoteClient = new MutableObject<>();
-
-        // flush producer in the middle of the cycle -- but we need a different thread to usePrev
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.removeRows(queryTable, i(10, 12));
-            TstUtils.addToTable(queryTable, i(5, 7), col("intCol", 10, 12));
-
-            final RowSetShiftData.Builder shiftBuilder = new RowSetShiftData.Builder();
-            shiftBuilder.shiftRange(0, 12, -5);
-
-            queryTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    shiftBuilder.build(), ModifiedColumnSet.EMPTY));
-
-            final BitSet cols = new BitSet(1);
-            cols.set(0);
-            remoteClient.setValue(
-                    remoteNugget.newClient(RowSetFactory.fromRange(0, 1), cols, "prevSnapshot"));
-
-            // flush producer in the middle of the cycle -- but we need a different thread to usePrev
-            final Thread thread = new Thread(this::flushProducerTable);
-            thread.start();
-            do {
-                try {
-                    thread.join();
-                } catch (final InterruptedException ignored) {
-
-                }
-            } while (thread.isAlive());
-        });
-
-        // We also have to flush the delta which is now in the pending list.
-        flushProducerTable();
-
-        // We expect two pending messages for our client: snapshot in prev and the shift update
-        Assert.equals(remoteClient.getValue().commandQueue.size(), "remoteClient.getValue().commandQueue.size()", 2);
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-
-        // validate
-        remoteNugget.validate("post flush");
-    }
-
-    public void testRegressModificationsInPrevView() {
-        final BitSet allColumns = new BitSet(1);
-        allColumns.set(0);
-
-        final QueryTable queryTable = TstUtils.testRefreshingTable(i(5, 10, 12).toTracking(),
-                col("intCol", 5, 10, 12));
-        final RemoteNugget remoteNugget = new RemoteNugget(() -> queryTable);
-
-        // Set original viewport.
-        final RemoteClient remoteClient =
-                remoteNugget.newClient(RowSetFactory.fromRange(1, 2), allColumns, "prevSnapshot");
-
-        // Obtain snapshot of original viewport.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("original viewport");
-
-        // Change viewport without overlap.
-        remoteClient.setViewport(RowSetFactory.fromRange(0, 1));
-
-        // Modify row that is outside of new viewport but in original.
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.addToTable(queryTable, i(12), col("intCol", 13));
-
-            queryTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    RowSetFactory.fromKeys(12),
-                    RowSetShiftData.EMPTY, ModifiedColumnSet.ALL));
-        });
-
-        // Do not allow the two updates to coalesce; we must force the consumer to apply the modification. (An allowed
-        // race.)
-        flushProducerTable();
-
-        // Add rows to shift modified row into new viewport.
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.removeRows(queryTable, i(5));
-
-            queryTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    RowSetFactory.fromKeys(5),
-                    RowSetFactory.empty(),
-                    RowSetShiftData.EMPTY, ModifiedColumnSet.EMPTY));
-        });
-
-        // Obtain snapshot of new viewport. (which will not include the modified row)
-        flushProducerTable();
-        Assert.equals(remoteClient.commandQueue.size(), "remoteClient.getValue().commandQueue.size()", 3); // mod, add,
-                                                                                                           // snaphot
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("new viewport with modification");
-    }
-
-    public void testCoalescingLargeUpdates() {
-        final BitSet allColumns = new BitSet(1);
-        allColumns.set(0);
-
-        final QueryTable sourceTable = TstUtils.testRefreshingTable(i().toTracking());
-        sourceTable.setFlat();
-        final QueryTable queryTable = (QueryTable) sourceTable.updateView("data = (short) k");
-
-        final RemoteNugget remoteNugget = new RemoteNugget(() -> queryTable);
-
-        // Create a few interesting clients around the mapping boundary.
-        final int mb = SNAPSHOT_CHUNK_SIZE;
-        final long sz = 2L * mb;
-        // noinspection unused
-        final RemoteClient[] remoteClients = new RemoteClient[] {
-                remoteNugget.newClient(null, allColumns, "full"),
-                remoteNugget.newClient(RowSetFactory.fromRange(0, 100), allColumns, "start"),
-                remoteNugget.newClient(RowSetFactory.fromRange(mb - 100, mb + 100), allColumns, "middle"),
-                remoteNugget.newClient(RowSetFactory.fromRange(sz - 100, sz + 100), allColumns, "end"),
-        };
-
-        // Obtain snapshot of original viewport.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("original viewport");
-
-        // Add all of our new rows spread over multiple deltas.
-        final int numDeltas = 4;
-        final long blockSize = sz / numDeltas;
-        for (int ii = 0; ii < numDeltas; ++ii) {
-            final RowSet newRows = RowSetFactory.fromRange(ii * blockSize, (ii + 1) * blockSize - 1);
-            updateGraph.runWithinUnitTestCycle(() -> {
-                TstUtils.addToTable(sourceTable, newRows);
-                sourceTable.notifyListeners(new TableUpdateImpl(
-                        newRows,
-                        RowSetFactory.empty(),
-                        RowSetFactory.empty(),
-                        RowSetShiftData.EMPTY, ModifiedColumnSet.ALL));
-            });
-        }
-
-        // Coalesce these to ensure mappings larger than a single chunk are handled correctly.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("large add rows update");
-
-        // Modify all of our rows spread over multiple deltas.
-        for (int ii = 0; ii < numDeltas; ++ii) {
-            final RowSetBuilderSequential modRowsBuilder = RowSetFactory.builderSequential();
-            for (int jj = ii; jj < sz; jj += numDeltas) {
-                modRowsBuilder.appendKey(jj);
-            }
-            updateGraph.runWithinUnitTestCycle(() -> {
-                sourceTable.notifyListeners(new TableUpdateImpl(
-                        RowSetFactory.empty(),
-                        RowSetFactory.empty(),
-                        modRowsBuilder.build(),
-                        RowSetShiftData.EMPTY, ModifiedColumnSet.ALL));
-            });
-        }
-
-        // Coalesce these to ensure mappings larger than a single chunk are handled correctly.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("large mod rows update");
-    }
-
-    /**
-     * Bug-fix verification: when a full subscription and a gapped-viewport subscription coexist, the producer stores
-     * all modified rows in a single delta (not just the viewport intersection). The viewport client's modOffsets then
-     * map contiguous viewport positions to non-contiguous data positions (with a gap). In appendModColumns, maxLength
-     * is computed as a data-position-space distance but then used as a row-index count, so modOffsets.get(endRange) can
-     * return a data position past the chunk boundary, triggering "Subset is out of bounds for context of size N".
-     */
-    public void testModColumnChunkBoundaryWithGappedViewport() {
-        final BitSet allColumns = new BitSet(1);
-        allColumns.set(0);
-
-        // Use a flat table with enough rows to span 2+ delta chunks when fully modified.
-        final long numRows = 2L * BarrageMessageProducer.DELTA_CHUNK_SIZE + 100;
-
-        final QueryTable sourceTable = TstUtils.testRefreshingTable(i().toTracking());
-        sourceTable.setFlat();
-        final QueryTable queryTable = (QueryTable) sourceTable.updateView("data = (short) k");
-
-        final RemoteNugget remoteNugget = new RemoteNugget(() -> queryTable);
-
-        // A full subscription is required so that modsToRecord = allRows (not just the viewport
-        // intersection). This makes rowsModified.original contiguous over all rows, which in turn
-        // makes clientModdedRowOffsets for the gapped-viewport client non-contiguous (gapped) —
-        // the precondition for the appendModColumns chunk-boundary bug.
-        // noinspection unused
-        final RemoteClient fullClient = remoteNugget.newClient(null, allColumns, "full");
-
-        // Create a viewport with a gap in the first chunk's range. The gap ensures that modOffsets
-        // is non-contiguous: viewport row indices 0..N map to data positions with a hole, so N row
-        // indices can map to data positions past the chunk boundary.
-        final long gapStart = BarrageMessageProducer.DELTA_CHUNK_SIZE / 2;
-        final long gapSize = BarrageMessageProducer.DELTA_CHUNK_SIZE / 4;
-        final RemoteClient remoteClient;
-        try (final RowSet before = RowSetFactory.fromRange(0, gapStart - 1);
-                final RowSet after = RowSetFactory.fromRange(gapStart + gapSize, numRows - 1)) {
-            final RowSet viewport = before.union(after);
-
-            // noinspection unused
-            remoteClient = remoteNugget.newClient(viewport, allColumns, "gapped-viewport");
-        }
-
-        // Obtain snapshot.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-
-        // Add all rows in one delta.
-        final RowSet allRows = RowSetFactory.fromRange(0, numRows - 1);
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.addToTable(sourceTable, allRows);
-            // This results in queryTable downstream update and the BMP propagating the adds.
-            sourceTable.notifyListeners(new TableUpdateImpl(
-                    allRows.copy(),
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    RowSetShiftData.EMPTY, ModifiedColumnSet.EMPTY));
-        });
-
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("after initial add");
-
-        // Modify ALL rows in a single delta. With the gapped viewport, the mod offsets will have
-        // a hole that causes appendModColumns to compute endPos past the first chunk's boundary.
-        updateGraph.runWithinUnitTestCycle(() -> {
-            // Faking a mods update from queryTable (since mods on a column-less table are impossible,
-            // doing this to trigger the BMP to reproduce the bug).
-            queryTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    allRows.copy(),
-                    RowSetShiftData.EMPTY, ModifiedColumnSet.ALL));
-        });
-
-        // This flush serializes the mod message; without the fix it throws:
-        // "Subset {..} is out of bounds for context of size N"
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("mod spanning chunks with gapped viewport");
-    }
-
-    public void testAllUniqueChunkTypeColumnSourcesWithValidityBuffers() {
-        testAllUniqueChunkTypeColumnSources(false);
-    }
-
-    public void testAllUniqueChunkTypeColumnSourcesWithDeephavenNulls() {
-        testAllUniqueChunkTypeColumnSources(true);
-    }
-
-    private void testAllUniqueChunkTypeColumnSources(final boolean useDeephavenNulls) {
-        this.useDeephavenNulls = useDeephavenNulls;
-
-        final int MAX_STEPS = 100;
-        for (int size : new int[] {10, 1000, 10000}) {
-            SharedProducerForAllClients helper =
-                    new SharedProducerForAllClients(1, 1, size, 0, new MutableInt(MAX_STEPS)) {
-                        @Override
-                        public void createTable() {
-                            // note we use column name `Sym` instead of `objCol` for the groupBy op in #createNuggets
-                            columnInfo = initColumnInfos(
-                                    new String[] {"longCol", "intCol", "Sym", "byteCol", "doubleCol", "floatCol",
-                                            "shortCol", "charCol", "boolCol", "strArrCol", "datetimeCol",
-                                            "bytePrimArray", "intPrimArray"},
-                                    new SortedLongGenerator(0, Long.MAX_VALUE - 1),
-                                    new IntGenerator(10, 100, 0.1),
-                                    new SetGenerator<>("a", "b", "c", "d"), // covers object
-                                    new ByteGenerator((byte) 0, (byte) 127, 0.1),
-                                    new DoubleGenerator(100.1, 200.1, 0.1),
-                                    new FloatGenerator(100.1f, 200.1f, 0.1),
-                                    new ShortGenerator((short) 0, (short) 20000, 0.1),
-                                    new CharGenerator('a', 'z', 0.1),
-                                    new BooleanGenerator(0.2),
-                                    new SetGenerator<>(new String[] {"a", "b"}, new String[] {"0", "1"},
-                                            new String[] {}, null),
-                                    new UnsortedInstantGenerator(
-                                            DateTimeUtils.parseInstant("2020-02-14T00:00:00 NY"),
-                                            DateTimeUtils.parseInstant("2020-02-25T00:00:00 NY")),
-                                    // uses var binary encoding
-                                    new ByteArrayGenerator(Byte.MIN_VALUE, Byte.MAX_VALUE, 0, 32),
-                                    // uses var list encoding
-                                    new IntArrayGenerator(0, Integer.MAX_VALUE, 0, 32));
-                            sourceTable = getTable(size / 4, random, columnInfo);
-                        }
-                    };
-
-            final int maxSteps = MAX_STEPS * helper.numConsumerCoalesce * helper.numProducerCoalesce;
-            helper.runTest(() -> {
-                final long lastKey = (Math.abs(helper.random.nextLong()) % 16)
-                        + (helper.sourceTable.isEmpty() ? -1 : helper.sourceTable.getRowSet().lastRowKey());
-                final TableUpdateImpl update = new TableUpdateImpl();
-                final int stepSize = Math.max(1, helper.size / maxSteps);
-                update.added = RowSetFactory.fromRange(lastKey + 1, lastKey + stepSize);
-                update.removed = i();
-                if (helper.sourceTable.isEmpty()) {
-                    update.modified = i();
-                } else {
-                    update.modified =
-                            RowSetFactory.fromRange(Math.max(0, lastKey - stepSize), lastKey);
-                    update.modified().writableCast().retain(helper.sourceTable.getRowSet());
-                }
-                update.shifted = RowSetShiftData.EMPTY;
-                update.modifiedColumnSet = ModifiedColumnSet.EMPTY;
-
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateTableUpdates(update,
-                        helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        }
-    }
-
-    public void testAllUniqueNonJavaSerRoundTripTypesWithValidityBuffers() {
-        testAllUniqueNonJavaSerRoundTripTypes(false);
-    }
-
-    public void testAllUniqueNonJavaSerRoundTripTypesWithDeephavenNulls() {
-        testAllUniqueNonJavaSerRoundTripTypes(true);
-    }
-
-    private void testAllUniqueNonJavaSerRoundTripTypes(final boolean useDeephavenNulls) {
-        this.useDeephavenNulls = useDeephavenNulls;
-
-        final int MAX_STEPS = 100;
-        for (int size : new int[] {10, 1000, 10000}) {
-            SharedProducerForAllClients helper =
-                    new SharedProducerForAllClients(1, 1, size, 0, new MutableInt(MAX_STEPS)) {
-                        @Override
-                        public void createTable() {
-                            // note we use column name `Sym` instead of `objCol` for the groupBy op in #createNuggets
-                            columnInfo = initColumnInfos(
-                                    new String[] {"longCol", "intCol", "Sym", "byteCol", "doubleCol", "floatCol",
-                                            "shortCol", "charCol", "boolCol", "strCol", "strArrCol", "datetimeCol",
-                                            "zdtCol", "localTimeCol", "localDateCol"},
-                                    new SortedLongGenerator(0, Long.MAX_VALUE - 1),
-                                    new IntGenerator(10, 100, 0.1),
-                                    new SetGenerator<>("a", "b", "c", "d"), // covers strings
-                                    new ByteGenerator((byte) 0, (byte) 127, 0.1),
-                                    new DoubleGenerator(100.1, 200.1, 0.1),
-                                    new FloatGenerator(100.1f, 200.1f, 0.1),
-                                    new ShortGenerator((short) 0, (short) 20000, 0.1),
-                                    new CharGenerator('a', 'z', 0.1),
-                                    new BooleanGenerator(0.2),
-                                    new StringGenerator(),
-                                    new SetGenerator<>(new String[] {"a", "b"}, new String[] {"0", "1"},
-                                            new String[] {}, null),
-                                    new UnsortedInstantGenerator(
-                                            DateTimeUtils.parseInstant("2020-02-14T00:00:00 NY"),
-                                            DateTimeUtils.parseInstant("2020-02-25T00:00:00 NY")),
-                                    new SetGenerator<>(
-                                            DateTimeUtils.parseInstant("2025-11-13T00:00:00 NY")
-                                                    .atZone(ZoneId.of("UTC")),
-                                            DateTimeUtils.parseInstant("2025-11-14T00:00:00 NY")
-                                                    .atZone(ZoneId.of("UTC")),
-                                            null),
-                                    new SetGenerator<>(
-                                            LocalTime.of(10, 30, 45),
-                                            LocalTime.of(14, 15, 30),
-                                            LocalTime.of(22, 45, 0),
-                                            null),
-                                    new SetGenerator<>(
-                                            LocalDate.of(2025, 11, 13),
-                                            LocalDate.of(2025, 11, 14),
-                                            LocalDate.of(2025, 11, 15),
-                                            null));
-                            sourceTable = getTable(size / 4, random, columnInfo);
-                        }
-                    };
-
-            final int maxSteps = MAX_STEPS * helper.numConsumerCoalesce * helper.numProducerCoalesce;
-            helper.runTest(() -> {
-                final long lastKey = (Math.abs(helper.random.nextLong()) % 16)
-                        + (helper.sourceTable.isEmpty() ? -1 : helper.sourceTable.getRowSet().lastRowKey());
-                final TableUpdateImpl update = new TableUpdateImpl();
-                final int stepSize = Math.max(1, helper.size / maxSteps);
-                update.added = RowSetFactory.fromRange(lastKey + 1, lastKey + stepSize);
-                update.removed = i();
-                if (helper.sourceTable.isEmpty()) {
-                    update.modified = i();
-                } else {
-                    update.modified =
-                            RowSetFactory.fromRange(Math.max(0, lastKey - stepSize), lastKey);
-                    update.modified().writableCast().retain(helper.sourceTable.getRowSet());
-                }
-                update.shifted = RowSetShiftData.EMPTY;
-                update.modifiedColumnSet = ModifiedColumnSet.EMPTY;
-
-                final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-                updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateTableUpdates(update,
-                        helper.random, helper.sourceTable, helper.columnInfo));
-            });
-        }
-    }
-
-    public void testVectorConcurrentModification() {
-        // this is a regression test for DH-19238; Barrage was not creating a static copy of ColumnWrapped vectors
-        final BitSet allColumns = new BitSet(2);
-        allColumns.set(0, 2);
-
-        final QueryTable queryTable = TstUtils.testRefreshingTable(RowSetFactory.flat(4).toTracking(),
-                col("Sym", "ZVZZT", "ZVZZT", "ZVZZT", "ZVZZT"),
-                col("Val", 1, 1, 1, 1));
-        final RemoteNugget remoteNugget = new RemoteNugget(() -> queryTable.groupBy("Sym"));
-
-        final RemoteClient remoteClient = remoteNugget.newClient(null, allColumns, "client");
-
-        // Obtain snapshot of original table.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("original viewport");
-
-        // Let's assert that our column source is virtual (important that the test is capable of failing).
-        Assert.eq(remoteNugget.originalTable.getColumnSource("Val").get(0).getClass(),
-                "remoteNugget.originalTable.getColumnSource(\"Val\").get(0).getClass()",
-                IntVectorColumnWrapper.class);
-
-        // Queue up a change to make Val == 2
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.addToTable(queryTable, queryTable.getRowSet(),
-                    col("Sym", "ZVZZT", "ZVZZT", "ZVZZT", "ZVZZT"),
-                    col("Val", 2, 2, 2, 2));
-
-            queryTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    queryTable.getRowSet().copy(),
-                    RowSetShiftData.EMPTY, ModifiedColumnSet.ALL));
-        });
-
-        // Queue up another change, but flush BMP before the table ticks. Client should receive 2's not 3's.
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.addToTable(queryTable, queryTable.getRowSet(),
-                    col("Sym", "ZVZZT", "ZVZZT", "ZVZZT", "ZVZZT"),
-                    col("Val", 3, 3, 3, 3));
-
-            flushProducerTable();
-
-            queryTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    RowSetFactory.empty(),
-                    queryTable.getRowSet().copy(),
-                    RowSetShiftData.EMPTY, ModifiedColumnSet.ALL));
-        });
-
-        // we're expecting a single update in the queue, flush and propagate to BarrageTable
-        Assert.equals(remoteClient.commandQueue.size(), "remoteClient.getValue().commandQueue.size()", 1);
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-
-        IntVector destVal = (IntVector) remoteClient.barrageTable.getColumnSource("Val").get(0);
-        Assert.neqNull(destVal, "destVal");
-        Assert.eq(destVal.size(), "destVal.size()", queryTable.size(), "queryTable.size()");
-        for (int ii = 0; ii < destVal.size(); ++ii) {
-            Assert.eq(destVal.get(ii), "destVal.get(ii)", 2);
-        }
-
-        // flush bmp and client one last time so we properly cleanup the test's chunks
-        flushProducerTable();
-        Assert.equals(remoteClient.commandQueue.size(), "remoteClient.getValue().commandQueue.size()", 1);
-        remoteNugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("end-of-test tables should match");
-    }
-
-    public void testNestedArrays() {
-        // Two cases that aren't expected to work in the java client, since it doesn't have enough type info:
-        // * triply nested groupBy(), where we have ObjectVector of ObjectVector of some vector
-        // * double nested groupBy() where the inner type is not a primitive
-        final Table queryTable = TableTools.emptyTable(1)
-                // .update("triplevector_int=i", "triplevector_string=``").groupBy()
-                .update("doublevector_int=i"/* , "doublevector_string=``" */).groupBy()
-                .update("vector_int=i", "vector_string=``").groupBy()
-                .update("array_int=new int[]{i}", "doublearray_int=new int[][] {{i}}",
-                        "array_string=new String[]{``}",
-                        "doublearray_string=new String[][]{{``}}");
-        queryTable.setRefreshing(true);
-        final BitSet allColumns = new BitSet(queryTable.getColumnSourceMap().size());
-        for (int i = 0; i < queryTable.getColumnSourceMap().size(); i++) {
-            allColumns.set(i);
-        }
-        final RemoteNugget remoteNugget = new RemoteNugget(() -> queryTable);
-
-        final RemoteClient remoteClient =
-                remoteNugget.newClient(RowSetFactory.fromRange(0, 0), allColumns, "snapshot");
-        remoteClient.setTransform(t -> {
-            // Based on the column type, if we see an array turn it into a string. While this seems to mask the types
-            // that we're getting over the wire, it doesn't function correctly without those types having arrived, so we
-            // should still be correctly validating the table.
-            return (QueryTable) t.updateView(t.getDefinition().getColumnNameMap().entrySet().stream()
-                    .map(entry -> {
-                        String colName = entry.getKey();
-                        Class<?> type = entry.getValue().getDataType();
-                        if (type.isArray()) {
-                            return Selectable.of(
-                                    ColumnName.of(colName),
-                                    RawString
-                                            .of("io.deephaven.server.barrage.BarrageMessageRoundTripTest.arrayToString("
-                                                    + colName + ")"));
-                        }
-                        return null;
-                    })
-                    .filter(Objects::nonNull)
-                    .collect(Collectors.toList()));
-        });
-        // Obtain snapshot of original viewport.
-        flushProducerTable();
-        remoteNugget.flushClientEvents();
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        remoteNugget.validate("snapshot");
-    }
-
-    public void testFailingNestedVectors() {
-        // Same as testNestedArrays, except we verify that we get expected exceptions for cases where BarrageTable can't
-        // handle the type
-        final Table failingTable = TableTools.emptyTable(1)
-                .update("triplevector_int=i", "triplevector_string=``").groupBy()
-                .update("doublevector_string=``").groupBy()
-                .groupBy();
-
-        failingTable.setRefreshing(true);
-        for (int i = 0; i < failingTable.getColumnSourceMap().size(); i++) {
-            BitSet oneColumn = new BitSet(failingTable.getColumnSourceMap().size());
-            oneColumn.set(i);
-            final RemoteNugget remoteNugget = new RemoteNugget(() -> failingTable);
-
-            final RemoteClient remoteClient =
-                    remoteNugget.newClient(RowSetFactory.fromRange(0, 0), oneColumn, "snapshot");
-            flushProducerTable();
-            remoteNugget.flushClientEvents();
-            final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-
-            assertNotNull(remoteClient.dummyObserver.failure);
-            assertTrue(remoteClient.dummyObserver.failure.getCause() instanceof GrpcMarshallingException);
-            assertTrue(remoteClient.dummyObserver.failure.getCause().getCause() instanceof StatusRuntimeException);
-            StatusRuntimeException sre =
-                    (StatusRuntimeException) remoteClient.dummyObserver.failure.getCause().getCause();
-            assertTrue(sre.getMessage().contains("Column with type java.lang.Object cannot be serialized directly"));
-        }
-    }
-
-    // Helpers for testNestedArrays
-    @ScriptApi
-    public static String arrayToString(int[] arr) {
-        return Arrays.toString(arr);
-    }
-
-    @ScriptApi
-    public static String arrayToString(Object[] arr) {
-        return Arrays.deepToString(arr);
-    }
-
-    // ---- Dictionary-encoding ticking tests ----
-
-    /**
-     * Builds a pojo {@link Schema} based on the natural schema of {@code def} except that the named column is annotated
-     * with an Arrow {@link DictionaryEncoding} (Int32 index, dict id 0).
-     * <p>
-     * Accepts a {@link TableDefinition} rather than a live {@link Table} to avoid calling
-     * {@link Table#getAttributes()}, which would publish the attribute map and prevent subsequent modification of the
-     * attributes.
-     */
-    private static Schema buildDictEncodedSchema(final TableDefinition def, final String dictColumnName) {
-        // Build the natural schema using an empty attributes map so getAttributes() is never called on the live table.
-        final Schema natural = BarrageUtil.makeSchema(
-                BarrageUtil.DEFAULT_SNAPSHOT_OPTIONS, def, Map.of(), false);
-        final List<Field> fields = natural.getFields().stream().map(f -> {
-            if (!dictColumnName.equals(f.getName())) {
-                return f;
-            }
-            return new Field(f.getName(),
-                    new FieldType(f.isNullable(), f.getType(),
-                            new DictionaryEncoding(0L, false, new ArrowType.Int(32, true)),
-                            f.getMetadata()),
-                    f.getChildren());
-        }).collect(Collectors.toList());
-        return new Schema(fields, natural.getCustomMetadata());
-    }
-
-    /**
-     * Builds a pojo {@link Schema} based on the natural schema of {@code def} except that the named column is doubly
-     * encoded as {@code RunEndEncoded<Dictionary<...>>}: the parent is run-end encoded (Int32 run_ends) and its
-     * {@code values} child carries an Arrow {@link DictionaryEncoding} (Int32 index, dict id 0).
-     */
-    private static Schema buildReeDictEncodedSchema(final TableDefinition def, final String colName) {
-        final Schema natural = BarrageUtil.makeSchema(
-                BarrageUtil.DEFAULT_SNAPSHOT_OPTIONS, def, Map.of(), false);
-        final List<Field> fields = natural.getFields().stream().map(f -> {
-            if (!colName.equals(f.getName())) {
-                return f;
-            }
-            // values child: the original value field annotated with a dictionary encoding
-            final Field values = new Field("values",
-                    new FieldType(f.isNullable(), f.getType(),
-                            new DictionaryEncoding(0L, false, new ArrowType.Int(32, true)), f.getMetadata()),
-                    f.getChildren());
-            final Field runEnds = new Field("run_ends",
-                    new FieldType(false, new ArrowType.Int(32, true), null), Collections.emptyList());
-            return new Field(f.getName(),
-                    new FieldType(false, new ArrowType.RunEndEncoded(), null, f.getMetadata()),
-                    List.of(runEnds, values));
-        }).collect(Collectors.toList());
-        return new Schema(fields, natural.getCustomMetadata());
-    }
-
-    /**
-     * Creates a two-column refreshing QueryTable ({@code Sym} String + {@code intCol} int) and annotates {@code Sym} as
-     * a doubly-encoded {@code RunEndEncoded<Dictionary<...>>} column via {@link Table#BARRAGE_SCHEMA_ATTRIBUTE}. The
-     * low-cardinality {@code Sym} generator produces both runs (favoring REE) and few distinct values (favoring the
-     * dictionary), exercising the combined encoding.
-     */
-    private QueryTable makeReeDictTable(final int initialSize, final Random random,
-            final ColumnInfo<?, ?>[] columnInfoOut) {
-        final ColumnInfo<?, ?>[] columnInfo = initColumnInfos(
-                new String[] {"Sym", "intCol"},
-                new SetGenerator<>("a", "b", "c"),
-                new IntGenerator(0, 100));
-        System.arraycopy(columnInfo, 0, columnInfoOut, 0, columnInfo.length);
-        final QueryTable table = getTable(initialSize, random, columnInfo);
-        table.setAttribute(Table.BARRAGE_SCHEMA_ATTRIBUTE, buildReeDictEncodedSchema(table.getDefinition(), "Sym"));
-        return table;
-    }
-
-    /**
-     * Full ticking subscription over a doubly-encoded {@code RunEndEncoded<Dictionary<...>>} column. Across many steps
-     * this exercises the nested dictionary's delta batches (new distinct values shipped as append-only DictionaryBatch
-     * messages preceding each RecordBatch) and the reader's run-expansion of dictionary indices.
-     */
-    public void testReeDictionaryEncodedFullSubscriptionTicking() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(0);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeReeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(null, allCols, "full-ree-dict");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    /**
-     * Two full subscribers on the same producer share a single {@code DictionaryWriterRegistry}; the nested dictionary
-     * of a {@code RunEndEncoded<Dictionary<...>>} column must still round-trip for both.
-     */
-    public void testReeDictionaryEncodedSharedProducer() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(1);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeReeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(null, allCols, "full-ree-dict-1");
-        nugget.newClient(null, allCols, "full-ree-dict-2");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    /**
-     * Regression test: a subscription whose initial snapshot carries <em>zero rows</em> (e.g. a freshly created,
-     * still-empty ticking table) must still emit an initial {@code isDelta=false} {@code DictionaryBatch} defining the
-     * nested dictionary's id <em>before</em> the first {@code RecordBatch} that references it. Strict Arrow consumers
-     * (such as the C++ client) fail or hang when the first RecordBatch references an undefined dictionary id, even with
-     * zero rows.
-     *
-     * <p>
-     * The Java reader is deliberately lenient here — {@code DictionaryChunkReader} skips dictionary resolution entirely
-     * for a zero-row batch — so a plain round-trip cannot detect the defect. Instead this test asserts the wire-level
-     * invariant directly, using the per-message {@link MessageHeader} types recorded by {@link DummyObserver}. Without
-     * the empty-batch dictionary-registration threading in
-     * {@code ChunkWriter.getEmptyInputStream(options, dictionaryRegistry)} (and its callers), the initial flush
-     * contains no DictionaryBatch at all and this test fails.
-     */
-    public void testReeDictionaryEncodedEmptyInitialSnapshot() {
-        final int steps = 10;
-        final int size = 100;
-        final Random random = new Random(2);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeReeDictTable(0, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        final RemoteClient client = nugget.newClient(null, allCols, "ree-dict-empty-initial");
-
-        // Deliver the initial snapshot; the table has no rows yet.
-        flushProducerTable();
-
-        final List<Byte> headerTypes = client.dummyObserver.observedHeaderTypes;
-        final int firstDictionaryBatch = headerTypes.indexOf(MessageHeader.DictionaryBatch);
-        final int firstRecordBatch = headerTypes.indexOf(MessageHeader.RecordBatch);
-        assertTrue("initial snapshot must contain a RecordBatch", firstRecordBatch >= 0);
-        assertTrue("initial snapshot of a dictionary-encoded subscription must emit a DictionaryBatch defining the "
-                + "dictionary id before the first RecordBatch that references it, even when the snapshot has zero "
-                + "rows; strict Arrow consumers (e.g. the C++ client) fail or hang otherwise. Observed header types: "
-                + headerTypes,
-                firstDictionaryBatch >= 0 && firstDictionaryBatch < firstRecordBatch);
-
-        // The (lenient) Java replica should also be a valid empty table at this point.
-        nugget.flushClientEvents();
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        nugget.validate("initial empty snapshot");
-
-        // Now grow the table from empty and confirm the stream stays healthy end-to-end.
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    /**
-     * Builds a schema where {@code col1Name} and {@code col2Name} both carry {@link DictionaryEncoding} with the same
-     * dictionary id (0). Two columns sharing a single id means a single {@code DictionaryBatch} per update covers both.
-     */
-    private static Schema buildSharedDictEncodedSchema(
-            final TableDefinition def, final String col1Name, final String col2Name) {
-        final Schema natural = BarrageUtil.makeSchema(
-                BarrageUtil.DEFAULT_SNAPSHOT_OPTIONS, def, Map.of(), false);
-        final DictionaryEncoding sharedEncoding =
-                new DictionaryEncoding(0L, false, new ArrowType.Int(32, true));
-        final List<Field> fields = natural.getFields().stream().map(f -> {
-            if (!col1Name.equals(f.getName()) && !col2Name.equals(f.getName())) {
-                return f;
-            }
-            return new Field(f.getName(),
-                    new FieldType(f.isNullable(), f.getType(), sharedEncoding, f.getMetadata()),
-                    f.getChildren());
-        }).collect(Collectors.toList());
-        return new Schema(fields, natural.getCustomMetadata());
-    }
-
-    /**
-     * Creates a two-column refreshing QueryTable ({@code Sym} String + {@code intCol} int) and annotates {@code Sym}
-     * with a dictionary encoding via {@link Table#BARRAGE_SCHEMA_ATTRIBUTE}. Returns the column-info array so callers
-     * can drive incremental updates via {@link GenerateTableUpdates}.
-     */
-    private QueryTable makeDictTable(final int initialSize, final Random random,
-            final ColumnInfo<?, ?>[] columnInfoOut) {
-        final ColumnInfo<?, ?>[] columnInfo = initColumnInfos(
-                new String[] {"Sym", "intCol"},
-                new SetGenerator<>("a", "b", "c"),
-                new IntGenerator(0, 100));
-        System.arraycopy(columnInfo, 0, columnInfoOut, 0, columnInfo.length);
-        final QueryTable table = getTable(initialSize, random, columnInfo);
-        table.setAttribute(Table.BARRAGE_SCHEMA_ATTRIBUTE, buildDictEncodedSchema(table.getDefinition(), "Sym"));
-        return table;
-    }
-
-    /**
-     * Creates a three-column refreshing QueryTable ({@code Sym1} String + {@code Sym2} String + {@code intCol} int)
-     * where both string columns share dictionary id=0 via {@link #buildSharedDictEncodedSchema}.
-     */
-    private QueryTable makeSharedDictTable(final int initialSize, final Random random,
-            final ColumnInfo<?, ?>[] columnInfoOut) {
-        final ColumnInfo<?, ?>[] columnInfo = initColumnInfos(
-                new String[] {"Sym1", "Sym2", "intCol"},
-                new SetGenerator<>("a", "b", "c"),
-                new SetGenerator<>("x", "y", "z"),
-                new IntGenerator(0, 100));
-        System.arraycopy(columnInfo, 0, columnInfoOut, 0, columnInfo.length);
-        final QueryTable table = getTable(initialSize, random, columnInfo);
-        table.setAttribute(Table.BARRAGE_SCHEMA_ATTRIBUTE,
-                buildSharedDictEncodedSchema(table.getDefinition(), "Sym1", "Sym2"));
-        return table;
-    }
-
-    public void testDictionaryEncodedFullSubscriptionTicking() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(0);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(null, allCols, "full-dict");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    public void testDictionaryEncodedSharedDictionaryAcrossFullSubscribers() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(1);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        // Two full subscribers on the same producer share a single DictionaryWriterRegistry
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(null, allCols, "full-dict-1");
-        nugget.newClient(null, allCols, "full-dict-2");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    /**
-     * Verifies that two columns sharing the same Arrow dictionary id (id=0) are correctly encoded and decoded through a
-     * full ticking subscription. The shared {@link io.deephaven.extensions.barrage.chunk.DictionaryWriterRegistry} must
-     * emit exactly one {@code DictionaryBatch} per id per update even though two columns reference it, and both columns
-     * must decode to their correct values.
-     */
-    public void testDictionaryEncodedSharedIdAcrossColumns() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(5);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[3];
-        final QueryTable sourceTable = makeSharedDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(null, allCols, "shared-id-full");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    public void testDictionaryEncodedViewportSubscriptionTicking() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(2);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        // Viewport subscriber gets its own private DictionaryWriterRegistry
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(RowSetFactory.fromRange(0, size / 10), allCols, "viewport-dict");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    public void testDictionaryEncodedGrowingSubscription() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(3);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        // Start as viewport subscription.
-        nugget.newClient(RowSetFactory.fromRange(0, size / 10), allCols, "growing-dict");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-
-        // Run a few steps as viewport
-        for (int step = 0; step < 5; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("viewport step " + step);
-        }
-
-        // Add a full subscription after a few viewport steps.
-        nugget.newClient(null, allCols, "growing-dict-full");
-
-        // Run remaining steps with both clients active.
-        for (int step = 5; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("mixed step " + step);
-        }
-    }
-
-    /**
-     * Creates a two-column refreshing QueryTable with many distinct Sym values (one per row by construction) so the
-     * dictionary accumulates quickly, enabling reliable overflow testing.
-     */
-    private QueryTable makeDictTableWithManyValues(final int initialSize, final String[] symValues,
-            final ColumnInfo<?, ?>[] columnInfoOut) {
-        final ColumnInfo<?, ?>[] columnInfo = initColumnInfos(
-                new String[] {"Sym", "intCol"},
-                new SetGenerator<>(symValues),
-                new IntGenerator(0, 1000));
-        System.arraycopy(columnInfo, 0, columnInfoOut, 0, columnInfo.length);
-        final QueryTable table = getTable(initialSize, new Random(42), columnInfo);
-        table.setAttribute(Table.BARRAGE_SCHEMA_ATTRIBUTE, buildDictEncodedSchema(table.getDefinition(), "Sym"));
-        return table;
-    }
-
-    /**
-     * Verifies that when the cumulative dictionary size exceeds the live row count, the server resets the dictionary
-     * (emitting {@code isDelta=false}) so the client stays consistent. Tests both full and viewport subscriptions.
-     */
-    public void testDictionaryEncodedOverflowCompaction() {
-        // Use 50 distinct values so the dictionary fills up quickly.
-        final String[] symValues = new String[50];
-        for (int i = 0; i < symValues.length; i++) {
-            symValues[i] = "val" + i;
-        }
-
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        // Start large enough that all 50 values are likely represented in the initial snapshot.
-        final QueryTable sourceTable = makeDictTableWithManyValues(200, symValues, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-
-        // Subscribe both a full client and a viewport client.
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        nugget.newClient(null, allCols, "overflow-full");
-        nugget.newClient(RowSetFactory.fromRange(0, 9), allCols, "overflow-viewport");
-
-        // Flush the initial snapshot.
-        flushProducerTable();
-        nugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        nugget.validate("initial snapshot");
-
-        // Remove most rows: keep only the first 5 live rows. The dictionary still has ~50 entries,
-        // but the live row count drops to 5 — triggering overflow on the next propagation.
-        final RowSet rowsToKeep = RowSetFactory.fromRange(0, 4);
-        final RowSet rowsToRemove = sourceTable.getRowSet().minus(rowsToKeep);
-        updateGraph.runWithinUnitTestCycle(() -> {
-            TstUtils.removeRows(sourceTable, rowsToRemove);
-            sourceTable.notifyListeners(new TableUpdateImpl(
-                    RowSetFactory.empty(),
-                    rowsToRemove.copy(),
-                    RowSetFactory.empty(),
-                    RowSetShiftData.EMPTY,
-                    ModifiedColumnSet.EMPTY));
-        });
-        rowsToRemove.close();
-
-        // propagateToSubscribers detects overflow (dict.size()~50 > liveRowCount=5) and resets the shared state.
-        flushProducerTable();
-        nugget.flushClientEvents();
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        nugget.validate("after mass removal - overflow triggered");
-
-        // Run a few more normal updates to confirm the dictionary rebuilds and correctness is maintained.
-        final Random random = new Random(99);
-        for (int step = 0; step < 10; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, 20, random, sourceTable, columnInfo));
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("post-overflow step " + step);
-        }
-    }
-
-    /**
-     * Verifies that a viewport subscription with a dictionary-encoded column correctly updates its local
-     * {@link io.deephaven.extensions.barrage.chunk.LocalDictionaryWriterState} when the viewport is shifted to rows
-     * that contain previously-unseen dictionary values.
-     *
-     * <p>
-     * This exercises the code path in {@code BarrageMessageProducer#propagateSnapshotForSubscription} that is reached
-     * after {@link RemoteClient#setViewport} triggers a new snapshot: the snapshot uses the subscription's private
-     * {@link io.deephaven.extensions.barrage.chunk.DictionaryWriterRegistry}, which must emit a fresh
-     * {@code isDelta=false} DictionaryBatch for the new window of values.
-     */
-    public void testDictionaryEncodedViewportChange() {
-        final int steps = 20;
-        final int size = 100;
-        final Random random = new Random(4);
-        final ColumnInfo<?, ?>[] columnInfo = new ColumnInfo<?, ?>[2];
-        final QueryTable sourceTable = makeDictTable(size / 4, random, columnInfo);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        // Start with a narrow viewport at the head of the table.
-        final int vpSize = size / 10;
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        final RemoteClient client = nugget.newClient(RowSetFactory.fromRange(0, vpSize - 1), allCols, "vp-dict");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-
-        // Run a few steps with the initial viewport, then shift the viewport every few steps.
-        // Shifting into new rows forces the producer to snapshot those rows; the dictionary must
-        // include any values in the new window that the client has not previously seen.
-        long vpStart = 0;
-        for (int step = 0; step < steps; step++) {
-            updateGraph.runWithinUnitTestCycle(() -> GenerateTableUpdates.generateShiftAwareTableUpdates(
-                    GenerateTableUpdates.DEFAULT_PROFILE, size, random, sourceTable, columnInfo));
-
-            // Every 4 steps, shift the viewport forward by the viewport size so it lands on a
-            // completely disjoint set of rows — maximizing the chance of encountering new Sym values.
-            if (step > 0 && step % 4 == 0) {
-                vpStart = (vpStart + vpSize) % Math.max(1, sourceTable.size() - vpSize);
-                final long finalVpStart = vpStart;
-                client.setViewport(RowSetFactory.fromRange(finalVpStart, finalVpStart + vpSize - 1));
-            }
-
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("step " + step);
-        }
-    }
-
-    // ---- Growing full subscription tests ----
-
-    /**
-     * The server grows a large full subscription's initial snapshot over several rounds; every growth message
-     * re-declares the entire consistent key space as {@code added} while only {@code rowsIncluded} is incremental. A
-     * client that consumes those messages across more than one of its own update graph cycles must publish none of
-     * them: after the first cycle the table already contains the whole key space, so a second update would claim rows
-     * as added that are already present. That is an invalid update in general, and it trips the
-     * {@link Table#APPEND_ONLY_TABLE_ATTRIBUTE} invariant that
-     * {@link io.deephaven.engine.table.impl.BaseTable#notifyListeners} checks whether or not a listener is attached.
-     * <p>
-     * The source is then ticked so that the deltas following the growth are validated against the contents the
-     * replicated table was left holding.
-     */
-    public void testGrowingFullSubscriptionAppendOnly() {
-        // The server ships at most MIN_SNAPSHOT_CELL_COUNT cells per growth round, so a single-column table needs more
-        // rows than that for the subscription to be grown across multiple messages.
-        final int size = (int) (BarrageUtil.MIN_SNAPSHOT_CELL_COUNT * 3);
-        final int[] values = new int[size];
-        for (int ii = 0; ii < size; ++ii) {
-            values[ii] = ii;
-        }
-        final QueryTable sourceTable = TstUtils.testRefreshingTable(
-                RowSetFactory.flat(size).toTracking(), TableTools.intCol("intCol", values));
-        sourceTable.setAttribute(Table.APPEND_ONLY_TABLE_ATTRIBUTE, true);
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        final RemoteClient client = nugget.newClient(null, allCols, "growing-append-only");
-
-        // let the server produce every growth message for the initial snapshot
-        flushProducerTable();
-        assertTrue("expected a growing snapshot, but the server sent " + client.pendingMessageCount() + " message(s)",
-                client.pendingMessageCount() > 1);
-
-        // deliver them one at a time, running an update graph cycle in between as a live client would
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-        while (client.pendingMessageCount() > 0) {
-            client.flushEventsToReplicatedTable(1);
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        }
-
-        // a null server viewport is how the server acknowledges that the full subscription is satisfied
-        assertNull(client.barrageTable.getServerViewport());
-        nugget.validate("growing append-only full subscription");
-
-        // Tick the source now that the subscription is satisfied. The replicated table's update validator was attached
-        // by the validate() above, taking the grown contents as its initial state, so these deltas must line up with
-        // what the growth left behind.
-        for (int step = 0; step < 3; ++step) {
-            final int firstRow = size + step * 10;
-            updateGraph.runWithinUnitTestCycle(() -> {
-                final int[] newValues = new int[10];
-                for (int ii = 0; ii < newValues.length; ++ii) {
-                    newValues[ii] = firstRow + ii;
-                }
-                final RowSet added = RowSetFactory.fromRange(firstRow, firstRow + newValues.length - 1);
-                TstUtils.addToTable(sourceTable, added, TableTools.intCol("intCol", newValues));
-                sourceTable.notifyListeners(new TableUpdateImpl(added, RowSetFactory.empty(), RowSetFactory.empty(),
-                        RowSetShiftData.EMPTY, ModifiedColumnSet.EMPTY));
-            });
-            flushProducerTable();
-            nugget.flushClientEvents();
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-            nugget.validate("post-growth step " + step);
-        }
-    }
-
-    /**
-     * An incomplete full subscription publishes no updates for the rows it is being given, so a listener attached
-     * before the server acknowledges the subscription would silently never hear about them. Attaching one must fail
-     * instead.
-     */
-    public void testListenToIncompleteFullSubscription() {
-        final int size = (int) (BarrageUtil.MIN_SNAPSHOT_CELL_COUNT * 3);
-        final int[] values = new int[size];
-        for (int ii = 0; ii < size; ++ii) {
-            values[ii] = ii;
-        }
-        final QueryTable sourceTable = TstUtils.testRefreshingTable(
-                RowSetFactory.flat(size).toTracking(), TableTools.intCol("intCol", values));
-
-        final BitSet allCols = new BitSet();
-        allCols.set(0, sourceTable.numColumns());
-
-        final RemoteNugget nugget = new RemoteNugget(() -> sourceTable);
-        final RemoteClient client = nugget.newClient(null, allCols, "incomplete-listen");
-
-        final ControlledUpdateGraph updateGraph = ExecutionContext.getContext().getUpdateGraph().cast();
-
-        // Deliver only the first of the growth messages, leaving the subscription unsatisfied.
-        flushProducerTable();
-        assertTrue(client.pendingMessageCount() > 1);
-        client.flushEventsToReplicatedTable(1);
-        updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        assertNotNull(client.barrageTable.getServerViewport());
-
-        try {
-            client.barrageTable.addUpdateListener(new FailureListener("Listener on incomplete table"));
-            TestCase.fail("expected an IllegalStateException listening to an incomplete table");
-        } catch (final IllegalStateException expected) {
-            assertTrue(expected.getMessage(), expected.getMessage().contains("incomplete table"));
-        }
-
-        // Once the subscription is satisfied the table may be listened to as usual.
-        while (client.pendingMessageCount() > 0) {
-            client.flushEventsToReplicatedTable(1);
-            updateGraph.runWithinUnitTestCycle(updateSourceCombiner::run);
-        }
-        assertNull(client.barrageTable.getServerViewport());
-        nugget.validate("satisfied full subscription");
-    }
-
-    public static class DummyObserver implements StreamObserver<BarrageMessageWriter.MessageView> {
-        volatile boolean completed = false;
-        volatile Throwable failure = null;
-
-        private final BarrageDataMarshaller marshaller;
-        private final Queue<BarrageMessage> receivedCommands;
-        /**
-         * The Arrow {@link MessageHeader} type of every wire message received, in arrival order (e.g.
-         * {@link MessageHeader#DictionaryBatch}, {@link MessageHeader#RecordBatch}). Lets tests assert wire-level
-         * message-ordering invariants that a lenient reader would otherwise paper over.
-         */
-        final List<Byte> observedHeaderTypes = new ArrayList<>();
-
-        DummyObserver(final BarrageDataMarshaller marshaller, final Queue<BarrageMessage> receivedCommands) {
-            this.marshaller = marshaller;
-            this.receivedCommands = receivedCommands;
-        }
-
-        /**
-         * Extracts the flatbuffer {@link MessageHeader} type from one FlightData-framed wire message, mirroring the
-         * header parsing in {@link BarrageMessageReaderImpl}.
-         */
-        private static byte peekHeaderType(final byte[] buf, final int len) throws IOException {
-            final CodedInputStream decoder = CodedInputStream.newInstance(buf, 0, len);
-            for (int tag = decoder.readTag(); tag != 0; tag = decoder.readTag()) {
-                if (tag == BarrageProtoUtil.DATA_HEADER_TAG) {
-                    final int size = decoder.readRawVarint32();
-                    return Message.getRootAsMessage(ByteBuffer.wrap(decoder.readRawBytes(size))).headerType();
-                }
-                decoder.skipField(tag);
-            }
-            return MessageHeader.NONE;
-        }
-
-        @Override
-        public void onNext(final BarrageMessageWriter.MessageView messageView) {
-            try {
-                messageView.forEachStream(inputStream -> {
-                    try (final ExposedByteArrayOutputStream baos = new ExposedByteArrayOutputStream()) {
-                        inputStream.drainTo(baos);
-                        inputStream.close();
-                        observedHeaderTypes.add(peekHeaderType(baos.peekBuffer(), baos.size()));
-                        final BarrageMessage message =
-                                marshaller.parse(new ByteArrayInputStream(baos.peekBuffer(), 0, baos.size()));
-                        // we skip schema messages, but can't suppress without propagating something...
-                        if (message != null) {
-                            receivedCommands.add(message);
-                        }
-                    } catch (final Exception e) {
-                        this.failure = e;
-                        throw new IllegalStateException("Failed to parse barrage message: ", e);
-                    }
-                });
-            } catch (final Exception e) {
-                this.failure = e;
-                throw new IllegalStateException("Failed to parse barrage message: ", e);
-            }
-        }
-
-        @Override
-        public void onError(final Throwable throwable) {
-            throw new IllegalStateException(throwable);
-        }
-
-        @Override
-        public void onCompleted() {
-            completed = true;
-        }
-    }
-}
